requirements.txt
fastapi
pymongo
pydantic_settings
uvicorn
google
google-genai
python-dotenv
streamlit
pandas
streamlit-agraph